backend/requirements.txt
fastapi
uvicorn
rq
redis
python-dotenv
boto3
reportlab
pillow
python-multipart
reportlab
python-docx
docx2pdf
pymupdf
